// src/components/shared/MachineForm/types.ts
export interface IFormInput {
  id: string;
  name: string;
  last10m: string;
  last30m: string;
  last1h: string;
  last1d: string;
  prev10m: string;
  prev30m: string;
  prev1h: string;
  prev1d: string;
  accepted: string;
  stale: string;
  dupelicate: string;
  other: string;
  contractId: string;
}